indicadores/requirements.txt
chardet==4.0.0
sqlalchemy==1.4.26
pandas==1.3.3
mysql-connector-python==8.0.27
